requirements.txt
Flask==0.10.1
Flask-JWT==0.3.2
Flask-PyMongo
python-dotenv==0.6.3
passlib
python-dateutil